Added google verification pages to build output, added about page, started on CMS previews for practices
