--- src/templates/about-page.js
@@ -1,91 +1,0 @@
-import React from "react";
-import { Box, Typography } from "@material-ui/core";
-import HeroColor from "../components/shared/HeroColor";
-import HistoryValues from "../components/About/HistoryValues";
-import ProductTeam from "../components/About/ProductTeam";
-import Feedback from "../components/About/Feedback";
-import Contributors from "../components/About/Contributors";
-
-export default function About(props) {
-  return (
-    <Box data-testid="aboutPageGrid" display="flex" flexDirection="column">
-      <HeroColor type="gradient" gradient={1}>
-        <Box px={6}>
-          <Typography component="h1" variant="h3" data-testid="hellotext">
-            About Us
-          </Typography>
-        </Box>
-      </HeroColor>
-      <Box maxWidth="70rem" alignSelf="center" mx={6}>
-        <HistoryValues />
-        <ProductTeam />
-        <Contributors />
-      </Box>
-      <Feedback />
-    </Box>
-  );
-}
-
-// import React from 'react'
-// import PropTypes from 'prop-types'
-// import { graphql } from 'gatsby'
-// import Layout from '../components/Layout'
-// import Content, { HTMLContent } from '../components/Content'
-
-// export const AboutPageTemplate = ({ title, content, contentComponent }) => {
-  // const PageContent = contentComponent || Content
-
-  // return (
-    // <section className="section section--gradient">
-      // <div className="container">
-        // <div className="columns">
-          // <div className="column is-10 is-offset-1">
-            // <div className="section">
-              // <h2 className="title is-size-3 has-text-weight-bold is-bold-light">
-                // {title}
-              // </h2>
-              // <PageContent className="content" content={content} />
-            // </div>
-          // </div>
-        // </div>
-      // </div>
-    // </section>
-  // )
-// }
-
-// AboutPageTemplate.propTypes = {
-  // title: PropTypes.string.isRequired,
-  // content: PropTypes.string,
-  // contentComponent: PropTypes.func,
-// }
-
-// const AboutPage = ({ data }) => {
-  // const { markdownRemark: post } = data
-
-  // return (
-    // <Layout>
-      // <AboutPageTemplate
-        // contentComponent={HTMLContent}
-        // title={post.frontmatter.title}
-        // content={post.html}
-      // />
-    // </Layout>
-  // )
-// }
-
-// AboutPage.propTypes = {
-  // data: PropTypes.object.isRequired,
-// }
-
-// export default AboutPage
-
-// export const aboutPageQuery = graphql`
-  // query AboutPage($id: String!) {
-    // markdownRemark(id: { eq: $id }) {
-      // html
-      // frontmatter {
-        // title
-      // }
-    // }
-  // }
-// `
