+++ static/admin/config.yml
@@ -3,7 +3,7 @@ backend:
   repo: openpracticelibrary/openpracticelibrary
   branch: next
 
-# local_backend: true
+local_backend: true
 
 publish_mode: editorial_workflow
 media_folder: "static/images" # Media files will be stored in the repo under images/uploads
@@ -28,14 +28,6 @@ collections:
         name: "date"
         widget: "datetime"
       - {label: "Authors", hint: "Comma-separated list of LOWER CASE github usernames eg., springdo, rdebeasi, tdbeattie", name: "authors", widget: "list", default: []}
-      - {label: "Area of the Library Loop", hint: "Choose the position on the Library Loop where your practice lives", name: "area", widget: "select", options:
-          [{label: "Discovery - Why", value: "discovery-loop-why"},
-          {label: "Discovery - Outcomes", value: "discovery-loop-outcomes"},
-          {label: "Options", value: "options"},
-          {label: "Delivery - Deliver", value: "delivery-deliver"},
-          {label: "Delivery - Measure & Learn", value: "delivery-measure-and-learn"},
-          {label: "Foundation - Culture & Collaboration", value: "foundation-culture-and-collaboration"},
-          {label: "Foundation - Technical",  value: "foundation-technical"}]}
       - label: "Tags"
         hint: "Where does this practice live on the Library Loop?"
         name: tags
@@ -67,22 +59,12 @@ collections:
             value: "measure"
           - label: "Learn (Delivery Loop)"
             value: "learn"
-      - {label: "Related Perspectives", hint: "Comma-separated list of slugs e.g., domain-driven-design, product-ownership", name: "perspectives", widget: "list", default: [], required: false,  pattern: ["^[a-z0-9-,]+$", "Must contain lower case and dash"]}
       - label: "Cover Image"
         name: "icon"
         hint: "Image to be used on the home screen and cards for the practice"
         widget: "image"
         required: true
         default: "https://github.com/openpracticelibrary/opl-media/blob/master/images/Needs%20an%20Image.png?raw=true"
-      - {label: "Jumbotron image", hint: "Large full width image for under the titles", name: "jumbotron", widget: "image", required: false}
-      - {label: "Jumbotron alt text", hint: "Accessible description of the image for screen readers", name: "jumbotronAlt", widget: "string", required: false}
-      - {label: "No. of people required", hint: "How many people are needed for the practice eg 2+ or 10", name: "people", widget: "string", required: false}
-      - {label: "Time length", hint: "How long is needed to facilitate this practice", name: "time", widget: "string", required: false}
-      - {label: "Facilitation difficulty", hint: "How difficult is it to facilitate this practice", name: "difficulty", widget: "select", required: false, options:
-          [{label: "Easy", value: "easy"},
-          {label: "Moderate", value: "moderate"},
-          {label: "Hard", value: "hard"}]}
-      - {label: "Expected participants", hint: "Who should attend when running this practice? eg., The Team, The Product Owner, BAs", name: "participants", required: false, widget: "list", default: []}
       - label: 'What is this practice?'
         name: 'whatIs'
         widget: 'markdown'
@@ -121,7 +103,6 @@ collections:
           - label: "Description"
             name: "description"
             widget: "string"
-      - {label: "Body", name: "body", widget: "markdown", default: "## What is it? \n\n## Why use it? \n\n## Further Information\n\n", required: false}
   - name: "perspective" # Used in routes, e.g., /admin/collections/blog
     label: "Perspective" # Used in the UI
     publish: false
